requirements.txt
Django==3.2.9
django-cors-headers==3.10.0
django-heroku==0.3.1
django-heroku-db-utils==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docutils==0.16
Pygments==2.7.4
PyJWT==2.3.0
pypiwin32==223
python-dateutil==1.5
python-decouple==3.4
pytz==2021.3
pywin32==300
requests-random-user-agent==2020.12.20

